requirements.txt
PyQt5
Pillow
opencv-python
matplotlib
